Add Styles_Child_Notices as subtree. See http://go.brain.st/15ExgfV
Excerpt from http://go.brain.st/15ExgfV
http://blogs.atlassian.com/2013/05/alternatives-to-git-submodule-git-subtree/

**Adding the sub-project as a remote**

Adding the subtree as a remote allows us to refer to it in shorter form:

    git remote add -f styles-child-notices git@github.com:stylesplugin/styles-child-notices.git

Now we can add the subtree (as before), but now we can refer to the remote in short form:

    git subtree add --prefix classes/styles-child-notices styles-child-notices master --squash

The command to update the sub-project at a later date becomes:

    git fetch styles-child-notices master
    git subtree pull --prefix classes/styles-child-notices styles-child-notices master --squash

**Contributing back to upstream**

We can freely commit our fixes to the sub-project in our local working directory now.

When it’s time to contribute back to the upstream project we need to fork the project and add it as another remote:

    git remote add styles-child-notices git@github.com:stylesplugin/styles-child-notices.git

Now we can use the subtree push command like the following:

    git subtree push --prefix=classes/styles-child-notices/ durdn-vim-surround master

    git push using:  styles-child-notices master
    Counting objects: 5, done.
    Delta compression using up to 4 threads.
    Compressing objects: 100% (3/3), done.
    Writing objects: 100% (3/3), 308 bytes, done.
    Total 3 (delta 2), reused 0 (delta 0)
    To git@github.com:stylesplugin/styles-child-notices.git
      02199ea..dcacd4b  dcacd4b21fe51c9b5824370b3b224c440b3470cb -} master

After this we’re ready and we can open a pull-request to the maintainer of the package.